LayoutTests/media/track/track-webvtt-tc010-no-timings-expected.txt

Tests cue without timings are ignored.


*** Testing text track 0
EXPECTED (cues.length == '0') OK
END OF TEST
